v4 Vote

+++ src/main/java/com/patryk/application/services/UserPrinciple.java
@@ -15,7 +15,7 @@ import com.patryk.application.models.User;
 public class UserPrinciple implements UserDetails {
 	private static final long serialVersionUID = 1L;
 
-	private Long id;
+	private int id;
 
 	private String name;
 
@@ -28,7 +28,7 @@ public class UserPrinciple implements UserDetails {
 
 	private Collection<? extends GrantedAuthority> authorities;
 
-	public UserPrinciple(Long id, String name, String username, String email, String password,
+	public UserPrinciple(int id, String name, String username, String email, String password,
 			Collection<? extends GrantedAuthority> authorities) {
 		this.id = id;
 		this.name = name;
@@ -46,7 +46,7 @@ public class UserPrinciple implements UserDetails {
 				authorities);
 	}
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
